Sound stuff
Ok, so since you isolated the dae file, I am not sure how to make a
specific key correspond (click wise) to a sound. Is there also a way to
make it centered visually?

I came up with something where I created (rough images, we can sleek
them up) and recorded myself playing a scale audio (scale.m4a) and have
the audio correspond to note clicks.

As a next step, it would be great to have the keyboard key light up
when the image is clicked. This can teach people how to read music!